+++ adma_tools/adma_tools_cpp/src/gsda_server.cpp
@@ -92,7 +92,7 @@ void GSDAServer::updateLoop()
 
   //offset between UNIX and GNSS (in ms)
   unsigned long long offset_gps_unix = 315964800000;
-  uint32_t week_to_msec = 604800000;
+  unsigned long long week_to_msec = 604800000;
   unsigned long long timestamp;
 
   while(rclcpp::ok())
